Refactor list_length type from uint16_t to uint8_t

+++ src/list.h
@@ -6,6 +6,6 @@ void calc_bytes_frequency_in_the_file(uint32_t *freq, FILE *file);
 void add_new_node_to_list(list_node **list, tree_node *leaf);
 void transfer_bytes_to_list(uint32_t *freq, list_node **list);
 list_node *create_char_list(FILE *file);
-uint16_t get_list_length(list_node *aux);
+uint8_t get_list_length(list_node *aux);
 
 #endif
